+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * encoded all URL parameters in pirsch.js where it is required
 * modernized pirsch.js
 * added reference implementation for pirsch-events.js
+* added more screen classes
 * fixed sending UTM parameters using pirsch.js
 * updated dependencies